Add WithLogger and configure logger into remote sampler
With this change, it is now possible to inject a logger into the
jaeger remote sampler. This makes it much easier to debug certain
configuration issues causing the sampler to not work.

The stdlib log.Default() logger will be used by default. This is
a change of behavior from before where no logger was used by default.
Users can still disable logging by calling WithLogger(logr.Discard()),
but it seems to me that errors should be exposed to the user by
default.

the github.com/go-logr/logr family provides quite a few wrapper types
so that other standard loggers can be used with this interface.