replace WithErrorPrinter with WithLogger and embed logger into jaegersampler
With this change, it is now possible to inject a logger into the
jaeger remote sampler. This makes it much easier to debug certain
configuration issues causing the sampler to not work.

The stdlib log.Default() logger will be used by default. This is
a change of behavior from before where no logger was used by default.
Users can still disable logging by calling WithLogger(logr.Discard()),
but it seems to me that errors should be exposed to the user by
default.