Warn the user if they are using deprecated remoteSamplerURL format
We used to expect a remoteSamplerURL that has a `service={}` at
the end of it. After 5f96ae48c29cdd8a60c7feebe862782012d89cbe we
now use a different API and we cannot tolerate a `service={}`.
If a user upgraded from the deprecated barney.ci/util/otel library
to this module, they might not realize they need to update their
remote sampler template URL. With this chnage we make it more
obvious.

Note that the default implementation of logr.Logger.Error will just
log to stderr. However some implementations may actually panic. The
logger implementation may be overridden by the user.