feat: add SDK and host info as tracer provider attrs (#23)
* feat: add SDK and host info as tracer provider attrs

With the opentracing client, we'd get host info (hostname, IP) by
default. This adds an option to include the same when using the otel
client.

It sets the two "standard" SDK info tags:

 - TelemetrySDKNameKey
 - TelemetrySDKVersionKey

I've also included an example so that users can have some idea what to
do when first coming to the library.