compat: fix UberTraceContext extraction (#24)
When we were extracing the span context from an uber-trace-id header,
we didn't properly handle the case of flags=0. It was expecting, for
the flags, a hex encoded string, but in fact, the flags is just a base
16 integer field.

The correct conversion is to just use ParseInt with base 16.

The test cases are taken from topic manager examples using the jaeger
client with go-opentracing.